+++ Sources/EpoxyBars/BarModel/AnyBarModel.swift
@@ -39,6 +39,10 @@ extension AnyBarModel: BarModeling {
 
 extension AnyBarModel: DataIDProviding {}
 
+// MARK: DidSelectProviding
+
+extension AnyBarModel: DidSelectProviding {}
+
 // MARK: Diffable
 
 extension AnyBarModel: Diffable {